+++ src/constants.ts
@@ -8,7 +8,7 @@
 // VERSION & DEBUG FLAGS
 ///////////////////////////////////////////////////////////////////////////////
 
-export const VERSION = [0, 0, 3]; // [major, minor, patch]
+export const VERSION = [0, 0, 4]; // [major, minor, patch]
 export const debugJSPlayer = true;
 export const debugMode = true; // DISABLE BEFORE SHARING